Only show scheduled or ongoing events on project page
